Yoni - seperation of the data
